Rename Flow class to SessionState
