+++ src/index.ts
@@ -13,8 +13,8 @@ const defaultURL = process.env.HASURA_GRAPHQL_URL;
 
 export * from "./passport";
 export * from "./logic";
+export { SessionState } from "./session-state";
 export { ComponentType } from "./types";
-export { Flow } from "./flow";
 export type {
   NodeId,
   Edges,
